pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-codereview"
version = "1.0.1"
description = "基于 Claude 的多维度 AI 代码审查工具"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["code-review", "claude", "anthropic", "ai", "static-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
git = ["gitpython>=3.1.0"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "gitpython>=3.1.0",
]

[project.scripts]
codereview = "review:app"

[tool.setuptools]
py-modules = [
    "baseline", "config", "eval", "llm_client",
    "parser", "preprocessor", "prompts", "review", "tools",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
